型と関数の命名を変更と移植、不要な型宣言を削除、送り出すデータのみにバリデーションをかけるように変更
