+++ nestJS/src/listing-item/listing-item.service.ts
@@ -1,11 +1,10 @@
 import { Injectable } from '@nestjs/common';
 import {
   ListingItemPayload,
-  RecievedListingItemPayload,
-  validateInput,
+  ListingItemInput,
 } from './interface/listing-item.payload';
 import { CustomPrismaService } from 'src/prisma/prisma.service';
-import { validate } from 'class-validator';
+import { validateValue } from 'src/common/validate-value';
 
 @Injectable()
 export class SearchItemService {
@@ -31,8 +30,8 @@ export class SearchItemService {
         firebaseUserId:true,
         imageUrls:true
       }});
-    const newitems = items.map((item) => {
-      const newitem:ListingItemPayload = {
+    const newItems = await Promise.all(items.map((item) => {
+      const newItem:ListingItemPayload = {
         id: item.id,
         createdAt: item.createdAt,
         documentId: item.documentId,
@@ -42,40 +41,30 @@ export class SearchItemService {
         price: item.price,
         name: item.name,
         firebaseUserId: item.firebaseUserId};
-      return newitem;
-    })
-    return newitems;
+      const checkedItem = validateValue(ListingItemPayload,newItem);  
+      return checkedItem;
+    }));
+    return newItems;
   }
 
-  async ItemCreate(createItem: RecievedListingItemPayload): Promise<string> {
-    try{
-      const validatedCreateItem = await validateInput(RecievedListingItemPayload,createItem)
-        .catch((error:Error)=>{
-          console.log(error); 
-          throw error;});
-      const adjustedCreateItem = {
-        //余計なプロパティの削除
-        documentId: validatedCreateItem.documentId,
-        condition: validatedCreateItem.condition,
-        department: validatedCreateItem.department,
-        price: validatedCreateItem.price,
-        name: validatedCreateItem.name,
-        firebaseUserId: validatedCreateItem.firebaseUserId,
-        imageUrls:{
-        create:validatedCreateItem.imageUrls.map((imageUrl)=>({url:imageUrl}))
-        } 
-      }
-      await this.prisma.listingItem.upsert({
-        where: { documentId: adjustedCreateItem.documentId },
-        update: adjustedCreateItem,
-        create:adjustedCreateItem,
-      });
-      return 'OK';
-
-    }catch(error){
-      if (error instanceof Error){
-        return error.message;
-      }
+  async ItemCreate(createItem: ListingItemInput): Promise<string> {
+    const adjustedCreateItem = {
+      //余計なプロパティの削除
+      documentId: createItem.documentId,
+      condition: createItem.condition,
+      department: createItem.department,
+      price: createItem.price,
+      name: createItem.name,
+      firebaseUserId: createItem.firebaseUserId,
+      imageUrls:{
+      create:createItem.imageUrls.map((imageUrl)=>({url:imageUrl}))
+      } 
     }
+    await this.prisma.listingItem.upsert({
+      where: { documentId: adjustedCreateItem.documentId },
+      update: adjustedCreateItem,
+      create:adjustedCreateItem,
+    });
+    return 'OK';
   }
 }
